+++ src/components/combo_box/combo_box.tsx
@@ -142,6 +142,15 @@ export interface _EuiComboBoxProps<T>
    * Specifies that the input should have focus when the component loads
    */
   autoFocus?: boolean;
+  /**
+   * Required when rendering without a visible label from [EuiFormRow](/#/forms/form-layouts).
+   */
+  'aria-label'?: string;
+  /**
+   * Reference ID of a text element containing the visible label for the combo box when not
+   * supplied by `aria-label` or from [EuiFormRow](/#/forms/form-layouts).
+   */
+  'aria-labelledby'?: string;
 }
 
 /**
@@ -924,6 +933,8 @@ export class EuiComboBox<T> extends Component<
       delimiter,
       append,
       autoFocus,
+      'aria-label': ariaLabel,
+      'aria-labelledby': ariaLabelledby,
       ...rest
     } = this.props;
     const {
@@ -1061,6 +1072,8 @@ export class EuiComboBox<T> extends Component<
           prepend={singleSelection ? prepend : undefined}
           isLoading={isLoading}
           autoFocus={autoFocus}
+          aria-label={ariaLabel}
+          aria-labelledby={ariaLabelledby}
         />
         {optionsList}
       </div>