+++ CHANGELOG.md
@@ -6,6 +6,7 @@ No public interface changes since `41.1.0`.
 
 - Added `layout` and `footer` props to `EuiEmptyPrompt` ([#5275](https://github.com/elastic/eui/pull/5275))
 - Updated `EuiEmptyPrompt` to extend `EuiPanelProps` ([#5275](https://github.com/elastic/eui/pull/5275))
+- Added `aria-label` and `aria-labelledby` props to `EuiComboBox` ([#5267](https://github.com/elastic/eui/issues/5267))
 - Add `data-icon-type` to `EuiIcon` `<svg>` for easier debugging of `iconType` [#5366](https://github.com/elastic/eui/pull/5366))
 
 **Bug fixes**